+++ src/org/infinity/resource/video/ImageRenderer.java
@@ -1,5 +1,5 @@
 // Near Infinity - An Infinity Engine Browser and Editor
-// Copyright (C) 2001 - 2005 Jon Olav Hauglid
+// Copyright (C) 2001 - 2022 Jon Olav Hauglid
 // See LICENSE.txt for license information
 
 package org.infinity.resource.video;
@@ -20,11 +20,10 @@ import javax.swing.SwingConstants;
 import org.infinity.gui.RenderCanvas;
 
 /**
- * A component optimized to display rapidly changing graphics data (e.g. videos, animations),
- * providing support for scaling (with and without aspect ratio preservation) and filtering.
+ * A component optimized to display rapidly changing graphics data (e.g. videos, animations), providing support for
+ * scaling (with and without aspect ratio preservation) and filtering.
  */
-public class ImageRenderer extends JComponent implements VideoBuffer, ComponentListener, SwingConstants
-{
+public class ImageRenderer extends JComponent implements VideoBuffer, ComponentListener, SwingConstants {
   // interpolation types used in scaling
   public static final Object TYPE_NEAREST_NEIGHBOR  = RenderingHints.VALUE_INTERPOLATION_NEAREST_NEIGHBOR;
   public static final Object TYPE_BILINEAR          = RenderingHints.VALUE_INTERPOLATION_BILINEAR;
@@ -36,22 +35,21 @@ public class ImageRenderer extends JComponent implements VideoBuffer, ComponentL
   private boolean isAspect;
 
   /**
-   * Creates an empty component. Call {@link #createBuffer(int, int, int)} to properly make use
-   * of the component's features.
+   * Creates an empty component. Call {@link #createBuffer(int, int, int)} to properly make use of the component's
+   * features.
    */
-  public ImageRenderer()
-  {
+  public ImageRenderer() {
     this(0, 0, 0);
   }
 
   /**
    * Creates a buffer chain with the specified parameters.
+   *
    * @param numBuffers The number of buffers to create.
-   * @param width The new width of the buffer in pixels.
-   * @param height The new height of the buffer in pixels.
+   * @param width      The new width of the buffer in pixels.
+   * @param height     The new height of the buffer in pixels.
    */
-  public ImageRenderer(int numBuffers, int width, int height)
-  {
+  public ImageRenderer(int numBuffers, int width, int height) {
     setOpaque(false);
     setLayout(null);
     videoBuffer = new BasicVideoBuffer();
@@ -63,14 +61,13 @@ public class ImageRenderer extends JComponent implements VideoBuffer, ComponentL
 
   /**
    * Creates a new buffer chain with the specified dimensions. Old buffers will be discarded.
+   *
    * @param numBuffers The number of buffers to create.
-   * @param width The new width of the buffer in pixels.
-   * @param height The new height of the buffer in pixels.
-   * @return {@code true} if the buffer chain has been created successfully,
-   *         {@code false} otherwise.
+   * @param width      The new width of the buffer in pixels.
+   * @param height     The new height of the buffer in pixels.
+   * @return {@code true} if the buffer chain has been created successfully, {@code false} otherwise.
    */
-  public boolean createBuffer(int numBuffers, int width, int height)
-  {
+  public boolean createBuffer(int numBuffers, int width, int height) {
     if (numBuffers > 0 && width > 1 && height > 1) {
       boolean res = videoBuffer.create(numBuffers, width, height, false);
       if (res) {
@@ -85,19 +82,19 @@ public class ImageRenderer extends JComponent implements VideoBuffer, ComponentL
 
   /**
    * Gets the alignment of the component's content along the X axis.
+   *
    * @return One of the following constants defined in SwingConstants: LEFT, CENTER or RIGHT.
    */
-  public int getHorizontalAlignment()
-  {
+  public int getHorizontalAlignment() {
     return canvas.getHorizontalAlignment();
   }
 
   /**
    * Sets the alignment of the component's content along the X axis. The default property is CENTER.
+   *
    * @param alignment One of the following constants defined in SwingConstants: LEFT, CENTER or RIGHT.
    */
-  public void setHorizontalAlignment(int alignment)
-  {
+  public void setHorizontalAlignment(int alignment) {
     if (alignment != canvas.getHorizontalAlignment()) {
       canvas.setHorizontalAlignment(alignment);
       updateCanvasBounds();
@@ -106,19 +103,19 @@ public class ImageRenderer extends JComponent implements VideoBuffer, ComponentL
 
   /**
    * Gets the alignment of the component's content along the Y axis.
+   *
    * @return One of the following constants defined in SwingConstants: TOP, CENTER or BOTTOM.
    */
-  public int getVerticalAlignment()
-  {
+  public int getVerticalAlignment() {
     return canvas.getVerticalAlignment();
   }
 
   /**
    * Sets the alignment of the component's content along the Y axis. The default property is CENTER.
+   *
    * @param alignment One of the following constants defined in SwingConstants: TOP, CENTER or BOTTOM.
    */
-  public void setVerticalAlignment(int alignment)
-  {
+  public void setVerticalAlignment(int alignment) {
     if (alignment != canvas.getVerticalAlignment()) {
       canvas.setVerticalAlignment(alignment);
       updateCanvasBounds();
@@ -127,10 +124,10 @@ public class ImageRenderer extends JComponent implements VideoBuffer, ComponentL
 
   /**
    * Returns the width of the image buffers set in the constructor or {@code createBuffer()} method.
+   *
    * @return Image buffer's width in pixels.
    */
-  public int getBufferWidth()
-  {
+  public int getBufferWidth() {
     if (videoBuffer.frontBuffer() != null) {
       return videoBuffer.frontBuffer().getWidth(null);
     } else {
@@ -140,10 +137,10 @@ public class ImageRenderer extends JComponent implements VideoBuffer, ComponentL
 
   /**
    * Returns the height of the image buffers set in the constructor or {@code createBuffer()} method.
+   *
    * @return Image buffer's height in pixels.
    */
-  public int getBufferHeight()
-  {
+  public int getBufferHeight() {
     if (videoBuffer.frontBuffer() != null) {
       return videoBuffer.frontBuffer().getHeight(null);
     } else {
@@ -153,19 +150,19 @@ public class ImageRenderer extends JComponent implements VideoBuffer, ComponentL
 
   /**
    * Returns whether scaling has been activated.
+   *
    * @return {@code true} if scaling has been activated.
    */
-  public boolean getScalingEnabled()
-  {
+  public boolean getScalingEnabled() {
     return canvas.isScalingEnabled();
   }
 
   /**
    * Sets whether the image buffer's content should be drawn scaled.
+   *
    * @param enable {@code true} to enable scaling.
    */
-  public void setScalingEnabled(boolean enable)
-  {
+  public void setScalingEnabled(boolean enable) {
     if (enable != canvas.isScalingEnabled()) {
       canvas.setScalingEnabled(enable);
       updateDefaultSize();
@@ -175,40 +172,39 @@ public class ImageRenderer extends JComponent implements VideoBuffer, ComponentL
 
   /**
    * Returns the interpolation type used when scaling has been enabled.
+   *
    * @return The interpolation type.
    */
-  public Object getInterpolationType()
-  {
+  public Object getInterpolationType() {
     return canvas.getInterpolationType();
   }
 
   /**
-   * Specify the interpolation type used when scaling has been enabled.
-   * One of TYPE_NEAREST_NEIGHBOR, TYPE_BILINEAR and TYPE_BICUBIC (Default: TYPE_NEAREST_NEIGHBOR).
+   * Specify the interpolation type used when scaling has been enabled. One of TYPE_NEAREST_NEIGHBOR, TYPE_BILINEAR and
+   * TYPE_BICUBIC (Default: TYPE_NEAREST_NEIGHBOR).
+   *
    * @param interpolationType The new interpolation type to set.
    */
-  public void setInterpolationType(Object interpolationType)
-  {
+  public void setInterpolationType(Object interpolationType) {
     canvas.setInterpolationType(interpolationType);
   }
 
   /**
    * Returns whether the aspect ratio of the displayed image will be preserved.
-   * @return {@code true} if aspect ratio preservation has been enabled,
-   *         {@code false} otherwise.
+   *
+   * @return {@code true} if aspect ratio preservation has been enabled, {@code false} otherwise.
    */
-  public boolean getAspectRatioEnabled()
-  {
+  public boolean getAspectRatioEnabled() {
     return isAspect;
   }
 
   /**
-   * Sets whether the aspect ratio of the displayed graphics should be preserved. This flag is used
-   * in conjunction with with {@link #setScalingEnabled(boolean)}.
+   * Sets whether the aspect ratio of the displayed graphics should be preserved. This flag is used in conjunction with
+   * with {@link #setScalingEnabled(boolean)}.
+   *
    * @param enable Set to {@code true} if aspect ratio should be preserved.
    */
-  public void setAspectRatioEnabled(boolean enable)
-  {
+  public void setAspectRatioEnabled(boolean enable) {
     if (enable != isAspect) {
       isAspect = enable;
       if (canvas.isScalingEnabled()) {
@@ -218,21 +214,19 @@ public class ImageRenderer extends JComponent implements VideoBuffer, ComponentL
   }
 
   /**
-   * Updates the renderer to use the current front buffer to be displayed. The renderer will
-   * display a cached image otherwise.
+   * Updates the renderer to use the current front buffer to be displayed. The renderer will display a cached image
+   * otherwise.
    */
-  public void updateRenderer()
-  {
+  public void updateRenderer() {
     canvas.setImage(videoBuffer.frontBuffer());
   }
 
   /**
    * Removes old content from all buffers in the buffer chain.
    */
-  public void clearBuffers()
-  {
+  public void clearBuffers() {
     for (int i = 0; i < videoBuffer.bufferCount(); i++) {
-      BufferedImage img = (BufferedImage)videoBuffer.frontBuffer();
+      BufferedImage img = (BufferedImage) videoBuffer.frontBuffer();
       if (img != null) {
         Graphics2D g = img.createGraphics();
         g.setColor(Color.BLACK);
@@ -244,75 +238,64 @@ public class ImageRenderer extends JComponent implements VideoBuffer, ComponentL
     updateRenderer();
   }
 
-//--------------------- Begin Interface VideoBuffer ---------------------
+  // --------------------- Begin Interface VideoBuffer ---------------------
 
   @Override
-  public Image frontBuffer()
-  {
+  public Image frontBuffer() {
     return videoBuffer.frontBuffer();
   }
 
   @Override
-  public Image backBuffer()
-  {
+  public Image backBuffer() {
     return videoBuffer.backBuffer();
   }
 
   @Override
-  public synchronized void flipBuffers()
-  {
+  public synchronized void flipBuffers() {
     videoBuffer.flipBuffers();
   }
 
   @Override
-  public int bufferCount()
-  {
+  public int bufferCount() {
     return videoBuffer.bufferCount();
   }
 
   @Override
-  public void attachData(Object data)
-  {
+  public void attachData(Object data) {
     videoBuffer.attachData(data);
   }
 
   @Override
-  public Object fetchData()
-  {
+  public Object fetchData() {
     return videoBuffer.fetchData();
   }
 
-//--------------------- End Interface VideoBuffer ---------------------
+  // --------------------- End Interface VideoBuffer ---------------------
 
-//--------------------- Begin Interface ComponentListener ---------------------
+  // --------------------- Begin Interface ComponentListener ---------------------
 
   @Override
-  public void componentHidden(ComponentEvent event)
-  {
+  public void componentHidden(ComponentEvent event) {
   }
 
   @Override
-  public void componentMoved(ComponentEvent event)
-  {
+  public void componentMoved(ComponentEvent event) {
   }
 
   @Override
-  public void componentResized(ComponentEvent event)
-  {
+  public void componentResized(ComponentEvent event) {
     if (event.getSource() == this) {
       updateCanvasBounds();
     }
   }
 
   @Override
-  public void componentShown(ComponentEvent event)
-  {
+  public void componentShown(ComponentEvent event) {
   }
 
-//--------------------- End Interface ComponentListener ---------------------
+  // --------------------- End Interface ComponentListener ---------------------
 
-  private void updateDefaultSize()
-  {
+  private void updateDefaultSize() {
     if (getScalingEnabled()) {
       setPreferredSize(getMinimumSize());
     } else {
@@ -320,8 +303,7 @@ public class ImageRenderer extends JComponent implements VideoBuffer, ComponentL
     }
   }
 
-  private void updateCanvasBounds()
-  {
+  private void updateCanvasBounds() {
     if (getWidth() > 0 && getHeight() > 0) {
       Rectangle newBounds = new Rectangle();
 
@@ -330,13 +312,13 @@ public class ImageRenderer extends JComponent implements VideoBuffer, ComponentL
         newBounds.width = getWidth();
         newBounds.height = getHeight();
         if (isAspect && getBufferHeight() > 0) {
-          float ar = (float)getBufferWidth() / (float)getBufferHeight();
-          if ((float)newBounds.width / ar <= newBounds.height) {
+          float ar = (float) getBufferWidth() / (float) getBufferHeight();
+          if (newBounds.width / ar <= newBounds.height) {
             // use width as base
-            newBounds.height = (int)((float)newBounds.width / ar);
+            newBounds.height = (int) (newBounds.width / ar);
           } else {
             // use height as base
-            newBounds.width = (int)((float)newBounds.height * ar);
+            newBounds.width = (int) (newBounds.height * ar);
           }
         }
       } else {
